tests cleaned up and still pass
